fix(colorWidgets): fix narrowing conversion

+++ src/gui/qgscolorwidgets.cpp
@@ -843,7 +843,7 @@ QgsColorBox::QgsColorBox( QWidget *parent, const ColorComponent component )
   setFocusPolicy( Qt::StrongFocus );
   setSizePolicy( QSizePolicy::MinimumExpanding, QSizePolicy::MinimumExpanding );
 
-  mBoxImage = std::make_unique<QImage>( static_cast<float>( width() ) - mMargin * 2, static_cast<float>( height() ) - mMargin * 2, QImage::Format_RGB32 );
+  mBoxImage = std::make_unique<QImage>( width() - static_cast<int>( mMargin * 2 ), height() - static_cast<int>( mMargin * 2 ), QImage::Format_RGB32 );
 }
 
 QgsColorBox::~QgsColorBox()
@@ -913,7 +913,7 @@ void QgsColorBox::setColor( const QColor &color, const bool emitSignals )
 void QgsColorBox::resizeEvent( QResizeEvent *event )
 {
   mDirty = true;
-  mBoxImage = std::make_unique<QImage>( static_cast<float>( event->size().width() ) - mMargin * 2, static_cast<float>( event->size().height() ) - mMargin * 2, QImage::Format_RGB32 );
+  mBoxImage = std::make_unique<QImage>( event->size().width() - static_cast<int>( mMargin * 2 ), event->size().height() - static_cast<int>( mMargin * 2 ), QImage::Format_RGB32 );
 
   QgsColorWidget::resizeEvent( event );
 }
